+++ src/bin/evebox.rs
@@ -313,6 +313,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         .subcommand(evebox::cli::sqlite::command())
         .subcommand(evebox::cli::test::command())
         .subcommand(evebox::cli::update::args())
+        .subcommand(evebox::cli::checkupdate::args())
         .subcommand(evebox::cli::util::args());
     let matches = parser.clone().get_matches();
 
@@ -347,6 +348,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         Some(("sqlite", args)) => evebox::cli::sqlite::main(args).await,
         Some(("test", args)) => evebox::cli::test::main(args).await,
         Some(("update", args)) => evebox::cli::update::main(args).await,
+        Some(("check-update", args)) => evebox::cli::checkupdate::main(args).await,
         Some(("util", args)) => evebox::cli::util::main(args).await,
         _ => {
             parser.print_help().ok();